TP/4/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
bitarray==3.2.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.1.0
click==8.1.8
cytoolz==1.0.1
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
Flask==3.1.0
frozenlist==1.5.0
hexbytes==1.3.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
multidict==6.2.0
packaging==24.2
parsimonious==0.10.0
pluggy==1.5.0
propcache==0.3.1
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
python-dotenv==1.1.0
pyunormalize==16.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
rpds-py==0.24.0
toolz==1.0.0
types-requests==2.32.0.20250306
typing_extensions==4.13.0
urllib3==2.3.0
web3==7.9.0
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
